+++ .vscode/settings.json
@@ -1,3 +1,20 @@
 {
-    "standard.nodePath": ""
+    "standard.enableGlobally": true,
+    "standard.usePackageJson": true,
+    "standard.engine": "ts-standard",
+    "standard.treatErrorsAsWarnings": false,
+    "standard.enable": true,
+    "standard.run": "onSave",
+    "standard.validate": [
+        "javascript",
+        "javascriptreact",
+        "typescript",
+        "typescriptreact"
+    ],
+    "standard.workingDirectories": [],
+    "standard.options": { },
+    "standard.nodePath": "${workspaceFolder}/app/node_modules",
+    "typescript.tsserver.log": "terse",
+    "js/ts.implicitProjectConfig.experimentalDecorators": false,
+    "js/ts.implicitProjectConfig.strictFunctionTypes": true,
 }
