--- app/BAKjest.config.ts
@@ -1,7 +1,0 @@
-export default {
-  clearMocks: true,
-  collectCoverage: true,
-  coverageDirectory: 'coverage',
-  preset: 'ts-jest',
-  testEnvironment: 'node'
-}
