+++ app/environment.d.ts
@@ -12,4 +12,4 @@ declare global {
 
 // If this file has no import/export statements (i.e. is a script)
 // convert it into a module by adding an empty export statement.
-export {}
+export { }
