+++ app/jest.config.ts
@@ -1,0 +1,7 @@
+import type { Config } from '@jest/types'
+
+// Sync object
+const config: Config.InitialOptions = {
+  verbose: true
+}
+export default config
