+++ app/package.json
@@ -6,16 +6,18 @@
   "scripts": {
     "build": "npx tsc",
     "prestart": "npm run build",
-    "start": "node dist/app.js",
+    "start": "node dist/index.js",
     "test": "jest --watchAll --no-cache --detectOpenHandles --runInBand --forceExit"
   },
   "ts-standard": {
-    "project": "./tsconfig.json",
+    "project": "./app/tsconfig.json",
     "ignore": [
       "dist",
       "node_modules",
       "coverage",
-      "app/**/*.js"
+      "app/**/*.js",
+      "${workspaceFolder}/app/environment.d.ts",
+      "${workspaceFolder}/app/jest.config.ts"
     ]
   },
   "keywords": [],
@@ -23,24 +25,28 @@
   "license": "ISC",
   "dependencies": {
     "cookie-parser": "^1.4.6",
-    "cookieparser": "^0.1.0",
     "dotenv": "^16.0.1",
-    "express": "^4.18.1"
+    "express": "^4.18.1",
+    "ts-standard": "^11.0.0"
   },
   "devDependencies": {
     "@babel/core": "^7.18.5",
     "@babel/preset-env": "^7.18.2",
     "@babel/preset-typescript": "^7.17.12",
+    "@types/cookie-parser": "^1.4.3",
     "@types/express": "^4.17.13",
     "@types/jest": "^28.1.3",
     "@types/node": "^18.0.0",
+    "@types/supertest": "^2.0.12",
+    "@typescript-eslint/eslint-plugin": "^5.29.0",
+    "@typescript-eslint/parser": "^5.29.0",
     "babel-jest": "^28.1.1",
+    "eslint": "^7.32.0",
     "jest": "^28.1.1",
     "standard": "^17.0.0",
     "supertest": "^6.2.3",
     "ts-jest": "^28.0.5",
     "ts-node": "^10.8.1",
-    "ts-standard": "^11.0.0",
     "typescript": "^4.7.4"
   }
 }
