+++ app/src/index.ts
@@ -1,0 +1,12 @@
+import app from './app'
+import * as dotenv from 'dotenv'
+import path from 'path'
+
+dotenv.config({ path: path.join(process.cwd(), '.env') })
+const port = parseInt(process.env.PORT ?? '3000', 10)
+
+app.listen(port, () => {
+  console.log(`⚡️[server]: Server is running at https://localhost:${port}`)
+})
+
+export default app
